+++ plugin.video.retrospect/resources/lib/helpers/languagehelper.py
@@ -1,6 +1,6 @@
 # SPDX-License-Identifier: CC-BY-NC-SA-4.0
 
-from addonsettings import AddonSettings
+from resources.lib.addonsettings import AddonSettings
 
 
 class LanguageHelper(object):
