+++ plugin.video.retrospect/resources/lib/helpers/xmlhelper.py
@@ -1,12 +1,12 @@
 # SPDX-License-Identifier: CC-BY-NC-SA-4.0
 
-from regexer import Regexer
-from helpers.taghelperbase import TagHelperBase
+from resources.lib.regexer import Regexer
+from resources.lib.helpers.taghelperbase import TagHelperBase
 
 #===============================================================================
 # Make global object available
 #===============================================================================
-from logger import Logger
+from resources.lib.logger import Logger
 
 
 class XmlHelper(TagHelperBase):
