+++ plugin.video.retrospect/resources/lib/streams/mms.py
@@ -1,8 +1,8 @@
 # SPDX-License-Identifier: CC-BY-NC-SA-4.0
 
-from regexer import Regexer
-from logger import Logger
-from urihandler import UriHandler
+from resources.lib.regexer import Regexer
+from resources.lib.logger import Logger
+from resources.lib.urihandler import UriHandler
 
 
 class Mms(object):
